Create GetSetImage.java
Image Processing In Java | Set 2 (Get and set Pixels)

In this set, we will learn about pixels of image. How we can get pixel values of image and how to set pixel values in an image using Java programming language. 
Pixels are the smallest unit of an image which consists of four components Alpha (transparency measure), Red, Green, Blue and in short (ARGB). 
The value of all the components lie between 0 and 255 both inclusive. Zero means the component is absent and 255 means the component is fully present.
Note: 
Since, 28 = 256 and the value of the pixel components lie between 0 and 255, so we need only 8-bits to store the values. 
SO, total number of bits required to store the ARGB values is 8*4=32 bits or 4 bytes. 
As the order signifies Alpha acquires leftmost 8 bits, Blue acquires rightmost 8 bits. 
Thus the bit position : 
For blue component being 7-0, 
For green component being 15-8, 
For red component being 23-16, 
For alpha component being 31-24.